Add view modifier to open a source code link
