--- Demo/Sources/Showcase/SourceCodeButton.swift
@@ -1,20 +1,0 @@
-//
-//  Copyright (c) SRG SSR. All rights reserved.
-//
-//  License information is available from the LICENSE file.
-//
-
-import SwiftUI
-
-struct SourceCodeButton: View {
-    let perform: () -> Void
-
-    var body: some View {
-        Button {
-            perform()
-        } label: {
-            Image(systemName: "globe")
-        }
-        .tint(.secondary)
-    }
-}
